refactor(spx-gui): refact select color function from fill component to select component
